Bug #1476519[PC-Code Scan][CameraIT-compare]加判断项目避免除数为0
[root cause  ] PC-lint Plus Code Scan
[changes     ] add if conditional statement
[side effects]no
[self test   ]:yes
[download normally]:yes
[power on/off normally]:yes
[do common repository/branch inspection]:yes
[is there dependence]:no
[confirm dependent commit]:no
[board]:sharkl5pro
[test case]:camera image compare
[reviewers   ]changdou.sun
[change_type ] feature_bugfix
[tag_product ] common

+++ test/framework/test_common_header.h
@@ -86,8 +86,8 @@ typedef enum {
 typedef struct compareInfo {
     unsigned char *test_img;
     unsigned char *golden_img;
-    int w;
-    int h;
+    int64_t w;
+    int64_t h;
     IT_IMG_FORMAT_T format;
 } compareInfo_t;
